requirements/prod.txt
channels==4.0.0
channels_redis==4.1.0
Django==4.2.7
django-cleanup==8.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
sorl-thumbnail==12.10.0
Pillow
